model-training/pyproject.toml
[project]
name = "model-training"
version = "0.1.0"
description = "Sub-repo for model training"
requires-python = ">=3.10,<4"
readme = "README.md"
dependencies = [
    "click>=8.2.0",
    "onnxruntime>=1.22.0",
    "onnxslim>=0.1.54",
    "pillow>=11.2.1",
    "poethepoet>=0.22.1,<0.23",
    "pydantic==2.12.2",
    "ultralytics>=8.3.136",
]

[project.scripts]
poe = "poethepoet:main"
model-train = "model_training.cli:train"

[project.entry-points."poetry.application.plugin"]
poethepoet = "poethepoet.plugin:PoetryPlugin"

[project.optional-dependencies]
labeling = [
    "label-studio>=1.20.0",
    "label-studio-converter>=0.0.59",
]

[dependency-groups]
ci = [
    "mypy>=1.5.1,<2",
    "black>=23.7.0,<24",
    "pylint>=2.17.5,<3",
    "isort>=5.12.0,<6",
    "pytest>=7.1.2,<8",
    "pytest-cov>=3.0.0,<4",
    "pycodestyle>=2.11.0,<3",
    "ruff>=0.11.8",
    "pre-commit",
    "yamllint>=1.37.1",
    "types-pyyaml>=6.0.12.20250516",
]
notebooks = [
    "jupyter>=1.1.1",
]

[tool.uv]
default-groups = ["ci"]

[tool.uv.sources]
ppq = { git = "https://github.com/espressif/esp-ppq.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["model_training"]

[tool.poe.tasks]
_clean_docs.script = "shutil:rmtree('docs/_build', ignore_errors=1)"

[tool.poe.tasks.train]
    help = "Runs an Ultralytics training job"
    cmd = "uv run python -m model_training.cli train"
    envfile = [".env", "local.env"]

[tool.poe.tasks.dev-setup]
    help = "Setup the project for development"
    cmd = "uv sync --all-groups"

[tool.poe.tasks.typing]
    help = "Run MyPy checker on the code base"
    cmd  = "mypy model_training"

[tool.poe.tasks.lint]
    help = "Run Ruff linter"
    cmd  = "ruff check"

[tool.poe.tasks.yamllint]
    help = "Run linter jobs for YAML files"
    cmd = "yamllint -f parsable configs/"

[tool.poe.tasks.lint-fix]
    help = "Fix linting and import errors"
    cmd = "ruff check --fix"

[tool.poe.tasks.fmt]
    help = "Format code using Ruff"
    cmd  = "ruff format"

[tool.poe.tasks.isort-fix]
    help = "Sort imports. This affects the code base as this makes changes to affected files."
    cmd = "isort model_training"

[tool.poe.tasks.isort]
    help = "Validate isort code style"
    cmd  = "isort model_training --check --diff"

[tool.poe.tasks.ci]
    help     = "Execute all CI tasks"
    sequence = ["typing", "lint", "yamllint", "isort"]
    ignore_fail = true

[tool.poe.tasks.clean]
    help = "Remove generated files"
    cmd  = """
    # multiline commands including comments work too!
    rm -rf .coverage
           .mypy_cache
           .pytest_cache
           ./**/__pycache__
           .ruff_cache
           dist
           htmlcov
           ./docs/_build
           ./tests/fixtures/simple_project/venv
           ./tests/fixtures/venv_project/myvenv
           ./tests/fixtures/poetry_plugin_project/**/.venv
           ./tests/temp
     """

[tool.poe.tasks.install-poetry-plugin]
    help = "Install or update this project as a plugin in poetry"
    sequence = [
      { cmd = "poetry self remove poethepoet"},
      { cmd = "poetry self add \"${POE_ROOT}[poetry_plugin]\""}
    ]
    ignore_fail = true

[tool.ruff]
line-length = 120
src = ["model_training"]

[tool.ruff.lint]
select = ["E", "F", "I"]
extend-select = ["I"]

[tool.isort]
profile = "black"
py_version = 310

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
plugins = ['pydantic.mypy']

[tool.pylint]
max-line-length = 120
disable = [
    "C0103",  # invalid-name
    "C0114",  # missing-module-docstring
    "W1203",  # logging-fstring-interpolation
    "E0213",  # no-self-argument
    "E1101",  # no-member
    "W0511",  # To-Dos
    "R0903"   # too-few-public-methods
]
